Agregada excepción ProjectException y cambiado tipo de dato de Long a Integer para proyectos

+++ src/main/java/com/uade/backendgestionbd2/controller/ProjectController.java
@@ -24,7 +24,7 @@ public class ProjectController {
 
 
     @GetMapping("/{userId}")
-    public List<Projects> getProjectsByUserId(@PathVariable Long userId) {
+    public List<Projects> getProjectsByUserId(@PathVariable int userId) {
         return projectService.getProjectsByUserId(userId);
     }
